SVY-18980 "Move Scope" contextual menu - update

+++ com.servoy.eclipse.model/src/com/servoy/eclipse/model/builder/ServoyBuilderUtils.java
@@ -212,7 +212,10 @@ public class ServoyBuilderUtils
 
 			try
 			{
-				file.deleteMarkers(ServoyBuilder.SCRIPT_MARKER_TYPE, true, IResource.DEPTH_INFINITE);
+				if (file.exists())
+				{
+					file.deleteMarkers(ServoyBuilder.SCRIPT_MARKER_TYPE, true, IResource.DEPTH_INFINITE);
+				}
 			}
 			catch (CoreException e)
 			{
